requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
idna==2.10
mysqlclient==2.0.1
pyTelegramBotAPI==3.7.3
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
